+++ app/pages/masters/construction-companies/[id].vue
@@ -1,0 +1,1 @@
+<template><MasterDetail master-type="constructionCompany" title="工務店マスター" /></template>