+++ app/pages/masters/homeowners/[id].vue
@@ -1,0 +1,1 @@
+<template><MasterDetail master-type="homeowner" title="施主マスター" /></template>