+++ app/pages/masters/properties/[id].vue
@@ -1,0 +1,1 @@
+<template><MasterDetail master-type="property" title="物件マスター" /></template>