+++ app/pages/masters/warranty-services/[id].vue
@@ -1,0 +1,1 @@
+<template><MasterDetail master-type="warrantyService" title="保証サービスマスター" /></template>